requirements.txt
# machine learning libraries
numpy
pandas
seaborn
matplotlib
scikit-learn
scipy

# file reading libraries
PyYAML

# ensure library
ensure

# box library
box

# google drive library
gdown

# dotenv
python-dotenv